requirements.txt
apache-airflow==2.4.3
pendulum==2.1.2
requests==2.26.0
kafka-python==2.0.2
Flask==2.2.5
Werkzeug==2.2.3
psycopg2-binary==2.9.6
mysql-connector-python
pandas
